wandb/run-20230723_093735-mxaehod4/files/requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.26.0
comm==0.1.2
conda-package-handling==1.7.3
conda==4.10.3
contourpy==1.0.7
cryptography==3.4.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
easydict==1.10
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
fonttools==4.39.0
fqdn==1.5.1
gitdb==4.0.10
gitpython==3.1.32
glcontext==2.4.0
glfw==1.12.0
google-auth-oauthlib==0.4.6
google-auth==2.16.2
grpcio==1.51.3
gym-notices==0.0.8
gym3==0.3.3
gym==0.26.2
idna==2.10
imageio-ffmpeg==0.3.0
imageio==2.31.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython-genutils==0.2.0
ipython==8.11.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
jinja2==3.1.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-server==2.4.0
jupyter-ydoc==0.2.3
jupyterlab-language-pack-zh-cn==3.6.post0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.5
jupyterlab==3.6.1
kiwisolver==1.4.4
lit==15.0.7
llvmlite==0.40.1
markdown==3.4.1
markupsafe==2.1.2
matplotlib-inline==0.1.6
matplotlib==3.7.1
mistune==2.0.5
moderngl==5.8.2
mpmath==1.3.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook-shim==0.2.2
notebook==6.5.3
numba==0.57.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==21.1.3
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
procgen==0.10.7
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.20
pygments==2.14.0
pyopenssl==20.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pysocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pyyaml==6.0
pyzmq==25.0.1
requests-oauthlib==1.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel-yaml-conda==0.15.100
send2trash==1.8.0
sentry-sdk==1.28.1
setproctitle==1.3.2
setuptools==52.0.0.post20210125
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
supervisor==4.2.5
sympy==1.11.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.12.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tornado==6.2
tqdm==4.61.2
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
wandb==0.15.5
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
werkzeug==2.2.3
wheel==0.36.2
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0